Several code improvements
